+++ Fun.Blazor/Store.fs
@@ -22,6 +22,15 @@ type IStoreManager =
     /// This is recommend way because you can use it with adaptiview easier.
     abstract CreateCVal: string * defautValue: 'T * init: (unit -> Task<'T>) -> cval<'T>
 
+    /// Create an adaptive list and share between components.
+    abstract CreateCList: string * ?defautValue: 'T seq -> clist<'T>
+
+    /// Create an adaptive map and share between components.
+    abstract CreateCHashSet: string * ?defautValue: 'T seq -> ChangeableHashSet<'T>
+
+    /// Create an adaptive map and share between components.
+    abstract CreateCMap: string * ?defautValue: ('K * 'T) seq -> cmap<'K, 'T>
+
     // Help us to access DI container
     abstract ServiceProvider: IServiceProvider
 
@@ -45,14 +54,14 @@ type IGlobalStore =
 
 
 type StoreManager(sp: IServiceProvider) =
-    let avals = ConcurrentDictionary<string, IAdaptiveValue>()
+    let avals = ConcurrentDictionary<string, obj>()
     let stores = ConcurrentDictionary<string, IDisposable>()
     let disposes = List<IDisposable>()
 
 
     interface IShareStore with
 
-        member _.CreateCVal(key: string, defaultValue: 'T) = avals.GetOrAdd(key, (fun _ -> cval defaultValue :> IAdaptiveValue)) :?> cval<'T>
+        member _.CreateCVal(key: string, defaultValue: 'T) = avals.GetOrAdd(key, (fun _ -> cval defaultValue)) :?> cval<'T>
 
         member _.CreateCVal(key: string, defaultValue: 'T, init: unit -> aval<'T>) =
             avals.GetOrAdd(
@@ -60,7 +69,7 @@ type StoreManager(sp: IServiceProvider) =
                 fun _ ->
                     let data = cval defaultValue
                     init().AddCallback(fun x -> transact (fun () -> data.Value <- x)) |> ignore
-                    data :> IAdaptiveValue
+                    data
             )
             :?> cval<'T>
 
@@ -70,9 +79,18 @@ type StoreManager(sp: IServiceProvider) =
                 fun _ ->
                     let data = cval defaultValue
                     init () |> Task.map (fun x -> transact (fun () -> data.Value <- x)) |> ignore
-                    data :> IAdaptiveValue
+                    data
             )
             :?> cval<'T>
+
+        member _.CreateCList(key: string, ?defaultValue: 'T seq) =
+            avals.GetOrAdd(key, fun _ -> clist (defaultArg defaultValue Seq.empty)) :?> clist<'T>
+
+        member _.CreateCHashSet(key: string, ?defaultValue: 'T seq) =
+            avals.GetOrAdd(key, fun _ -> ChangeableHashSet(defaultArg defaultValue Seq.empty)) :?> ChangeableHashSet<'T>
+
+        member _.CreateCMap(key: string, ?defaultValue: ('K * 'T) seq) =
+            avals.GetOrAdd(key, fun _ -> cmap<'K, 'T> (defaultArg defaultValue Seq.empty)) :?> cmap<'K, 'T>
 
         member _.ServiceProvider = sp