+++ Fun.Blazor/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+## [4.1.5] - 2025-02-28
+
+- Add CreateCList for store
+- Add CreateCHashSet for stroe
+- Add CreateCMap for store
+
 ## [4.1.4] - 2024-12-10
 
 - Fix createHiddenInputs bool property handling